import stdint.h

+++ mosken.h
@@ -1,6 +1,7 @@
 #ifndef MOSKEN_H
 #define MOSKEN_H
 
+#include <stdint.h>
 #include <stddef.h>
 #define BLKSZ 4096 // size of a db block in bytes
 
@@ -14,26 +15,26 @@ typedef struct
 {
     int payment_id;
     char payment_name[24];
-    unsigned long payment_time;
-    unsigned long total_cents;
+    uint32_t payment_time;
+    uint32_t total_cents;
 } PaymentData;
 
 typedef PaymentData *Payment;
 
-typedef unsigned short Offset;
+typedef uint32_t Offset;
 
 typedef struct
 {
     Offset pgi_offset; /* offset into the page to the start of the item */
-    unsigned short pgi_length; /* length of the item in the page */
+    uint32_t pgi_length; /* length of the item in the page */
 } PageItemMeta;
 
 typedef PageItemMeta *PgItemId;
 
 typedef struct
 {
-    unsigned short pgh_lower; /* offset to the start of free space in the page */
-    unsigned short pgh_upper; /* offset to the end of free space in the page */
+    uint32_t pgh_lower; /* offset to the start of free space in the page */
+    uint32_t pgh_upper; /* offset to the end of free space in the page */
     PageItemMeta pgh_im[];    /* item metadata array */
 } PgHeaderData;
 
@@ -58,9 +59,9 @@ typedef struct
 typedef PageDirectoryData *PageDirectory;
 
 typedef char *Page;
-typedef size_t Size;
+typedef uint32_t Size;
 
 typedef char *Item;
-typedef unsigned short OffsetNum;
+typedef uint32_t OffsetNum;
 
 #endif // !MOSKEN_H
